Consolidate sparse MLA CCL benchmarks into one file
Fold sparse_mla_ccl.py back into test_sparse_mla_ccl_perf.py so the benchmarks
live in a single file. Drop the design-doc references from the module and
function docstrings, and document the CCLTraffic.critical_path_bytes (latency
term: bytes on the longest single dependency chain) and total_network_bytes
(aggregate mesh traffic, not latency) fields.

Validated: all 6 tests pass on a LoudBox (8 Blackhole chips) under the tracy
realtime profiler.
